requirements.txt
Flask==1.0.2
requests==2.18.4
pycryptodome==3.8.0
ujson==1.35